(PATCH)feat: stuff meade with navbar
