+++ Frontend/src/Domain/ProfileStats.tsx
@@ -1,0 +1,13 @@
+﻿interface ProfileStats {
+    name: string;
+    timeOfRegistry: string;
+    totalAmountOfWorkouts: number;
+    totalAmountOfTimeTrained: string;
+    currentStreakDays: number;
+    currentStreakWeeks: number;
+    bestStreakDays: number;
+    bestStreakWeeks: number;
+    favoriteExercise: string;
+}
+
+export default ProfileStats;
