+++ Frontend/src/assets/Components/NavBar.css
@@ -12,3 +12,4 @@
     background-color: #333;
     padding: 0.5rem 1rem;
 }
+
